email-outreach/agents/user-context/requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
PyMuPDF==1.23.26
